+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 - Mechtild, the new image artist, is now available for image generation based on Dall-E 3.
+- Add shortcuts to the navigation for quick access to the images, documents, and conversations.
 
 ### Fixed
 - Resolved importing from alpha-0.2 does not assign all files and images to a directory.